Move framework definition loading to separate file
There was enough code to warrant its own module.
