+++ amlb/resources.py
@@ -3,16 +3,14 @@
 as well as handy methods to access other resources like *automl frameworks* and *benchmark definitions*
 """
 import copy
-import itertools
 import logging
 import os
-from queue import Queue
 import random
 import re
 import sys
-from typing import List, Union, Optional
 
 from amlb.benchmarks.parser import benchmark_load
+from amlb.framework_definitions import load_framework_definitions
 from .utils import Namespace, config_load, lazy_property, memoize, normalize_path, touch
 
 
@@ -192,180 +190,6 @@ class Resources:
         if task[conf] is None:
             task[conf] = self.config.aws.ec2.volume_type
             log.debug("Config `{config}` not set for task {name}, using default `{value}`.".format(config=conf, name=task.name, value=task[conf]))
-
-
-def load_and_merge_framework_definitions(frameworks_file: Union[str, List[str]]) -> Namespace:
-    """ Load and merge the framework file(s), does not allow duplicate definitions. """
-    log.info("Loading frameworks definitions from %s.", frameworks_file)
-    if not isinstance(frameworks_file, list):
-        frameworks_file = [frameworks_file]
-
-    definitions_by_file = [config_load(file) for file in frameworks_file]
-    for d1, d2 in itertools.combinations([set(dir(d)) for d in definitions_by_file], 2):
-        if d1.intersection(d2) != set():
-            raise ValueError(f"Duplicate entry '{d1.intersection(d2).pop()}' found.")
-    return Namespace.merge(*definitions_by_file)
-
-
-def sanitize_definitions(frameworks: Namespace):
-    """ Normalize names, add name field, remove invalid extensions. """
-    add_and_normalize_names(frameworks)
-    remove_frameworks_with_unknown_parent(frameworks)
-    remove_self_reference_extensions(frameworks)
-
-
-def sanitize_and_add_defaults(frameworks, resource: Resources):
-    sanitize_definitions(frameworks)
-
-    # `module` is the only field that should have a default
-    # based on the parent. For that reason we add it before
-    # we update children with their parent fields.
-    for _, framework in frameworks:
-        if "extends" not in framework:
-            add_default_module(framework, resource.config)
-    update_frameworks_with_parent_definitions(frameworks)
-
-    add_defaults_to_frameworks(frameworks, resource)
-
-
-def load_framework_definitions(frameworks_file: Union[str, List[str]], resource: Resources) -> Namespace:
-    """ Load the framework definition listed in the framework file(s).
-
-    Loads the definition(s) from the file(s),
-    :param frameworks_file:
-    :return: Namespace containing each framework definition,
-    """
-    frameworks = load_and_merge_framework_definitions(frameworks_file)
-    sanitize_and_add_defaults(frameworks, resource)
-    log.debug("Available framework definitions:\n%s", frameworks)
-    return frameworks
-
-
-def add_and_normalize_names(frameworks: Namespace):
-    """ Converts each framework definition to lowercase and adds a 'name' field. """
-    framework_names = dir(frameworks)
-    for name in framework_names:
-        framework = frameworks[name]
-        framework.name = name.lower()
-        if name.lower() != name:
-            del frameworks[name]
-            frameworks[name.lower()] = framework
-        if "extends" in framework:
-            framework.extends = framework.extends.lower()
-
-
-def add_default_module(framework, config):
-    if "module" not in framework:
-        framework.module = f"{config.frameworks.root_module}.{framework.name}"
-
-
-def add_default_version(framework):
-    if "version" not in framework:
-        framework.version = "latest"
-
-
-def add_default_setup_args(framework):
-    if "setup_args" in framework:
-        framework.setup_args = [framework.setup_args]
-    else:
-        framework.setup_args = [framework.version]
-        if "repo" in framework:
-            framework.setup_args.append(framework.repo)
-
-
-def add_default_setup_script(framework, resource):
-    if "setup_script" not in framework:
-        framework.setup_script = None
-    else:
-        framework.setup_script = framework.setup_script.format(
-            module=framework.module,
-            **resource._common_dirs,
-        )
-
-
-def add_default_setup_cmd(framework, resource):
-    if "setup_cmd" not in framework:
-        framework._setup_cmd = None
-        framework.setup_cmd = None
-    else:
-        framework._setup_cmd = framework.setup_cmd
-        if isinstance(framework.setup_cmd, str):
-            framework.setup_cmd = [framework.setup_cmd]
-        framework.setup_cmd = [
-            cmd.format(pip="{pip}", py="{py}", **resource._common_dirs)
-            for cmd in framework.setup_cmd
-        ]
-
-
-def add_default_params(framework):
-    if "params" not in framework:
-        framework.params = dict()
-    else:
-        framework.params = Namespace.dict(framework.params)
-
-
-def add_default_image(framework: Namespace, config_: Namespace):
-    if "image" not in framework:
-        framework.image = copy.deepcopy(config_.docker.image_defaults)
-    else:
-        framework.image = Namespace.merge(config_.docker.image_defaults, framework.image)
-
-    if framework.image.tag is None:
-        framework.image.tag = framework.version.lower()
-
-    if framework.image.image is None:
-        framework.image.image = framework.name
-
-
-def find_all_parents(framework, frameworks):
-    """ Return all definitions framework extends, from direct parent to furthest. """
-    parents = []
-    while "extends" in framework and framework.extends is not None:
-        framework = frameworks[framework.extends]
-        parents.append(framework)
-    return parents
-
-
-def update_frameworks_with_parent_definitions(frameworks: Namespace):
-    """ Add fields defined by ancestors
-
-    Extensions do not overwrite fields defined on the framework itself.
-    If multiple parents define the same field, the parent that is 'closer'
-    to the child framework defines the field value.
-    """
-    for name, framework in frameworks:
-        log.info(f"{name} {framework}")
-        parents = find_all_parents(framework, frameworks)
-        for parent in parents:
-            framework % copy.deepcopy(parent)
-
-
-def add_defaults_to_frameworks(frameworks: Namespace, resource: Resources):
-    for _, framework in frameworks:
-        add_default_version(framework)
-        add_default_setup_args(framework)
-        add_default_params(framework)
-        add_default_image(framework, resource.config)
-        add_default_setup_cmd(framework, resource)
-        add_default_setup_script(framework, resource)
-
-
-def remove_self_reference_extensions(frameworks: Namespace):
-    for name, framework in frameworks:
-        if "extends" in framework and framework.extends == framework.name:
-            log.warning("Framework %s extends itself: removing extension.",
-                        framework.name)
-            framework.extends = None
-
-
-def remove_frameworks_with_unknown_parent(frameworks: Namespace):
-    frameworks_with_unknown_parent = [
-        (name, framework.extends) for name, framework in frameworks
-        if "extends" in framework and framework.extends not in frameworks
-    ]
-    for framework, parent in frameworks_with_unknown_parent:
-        log.warning("Removing framework %s as parent %s doesn't exist.", framework, parent)
-        del frameworks[framework]
 
 
 __INSTANCE__: Resources = None
